Delete skills app, rename users app to users_skills app, create api app
