--- skills/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class SkillsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'skills'
-    verbose_name = 'Навыки'
