--- skills/constants.py
@@ -1,2 +1,0 @@
-MAX_LENGTH_KEY_SKILL = 4
-MAX_LENGTH_REQUIRED_LEVEL = 128
