+++ users_skills/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class UsersSkillsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'users_skills'
+    verbose_name = 'Пользователи_Навыки'
