+++ users_skills/constants.py
@@ -2,6 +2,8 @@ MAX_LENGTH_ABOUT_TEAM = 128
 MAX_LENGTH_FIRSTNAME = 64
 MAX_LENGTH_LASTNAME = 128
 MAX_LENGTH_KEY_EMPLOYEE = 4
+MAX_LENGTH_KEY_SKILL = 4
+MAX_LENGTH_REQUIRED_LEVEL = 128
 MAX_LENGTH_RIGHTS = 4
 MAX_LENGTH_ROLE = 64
 MAX_LENGTH_URL_CONFLUENCE = 128
